Breaking stuff out so I can write tests
