+++ build.gradle
@@ -19,6 +19,8 @@ dependencies {
   implementation 'commons-io:commons-io:2.6'
   implementation 'org.javassist:javassist:3.21.0-GA'
   implementation 'io.sentry:sentry:1.7.30'
+  testImplementation "org.assertj:assertj-core:3.8.0"
+  testImplementation "io.mockk:mockk:1.10.0"
 }
 
 configurations {
