+++ src/main/kotlin/io/unthrottled/doki/service/AppService.kt
@@ -1,0 +1,8 @@
+package io.unthrottled.doki.service
+
+import com.intellij.openapi.application.ApplicationNamesInfo
+
+object AppService {
+  fun getApplicationName(): String =
+    ApplicationNamesInfo.getInstance().fullProductNameWithEdition
+}
